+++ src/mavsdk/plugins/camera_server/include/plugins/camera_server/camera_server.h
@@ -841,12 +841,12 @@ public:
     /**
      * @brief Callback type for subscribe_tracking_off_command.
      */
-    using TrackingOffCommandCallback = std::function<void(int32_t)>;
+    using TrackingOffCommandCallback = std::function<void(uint64_t)>;
 
     /**
      * @brief Handle type for subscribe_tracking_off_command.
      */
-    using TrackingOffCommandHandle = Handle<int32_t>;
+    using TrackingOffCommandHandle = Handle<uint64_t>;
 
     /**
      * @brief Subscribe to incoming tracking off command.
